eeff_ctrl_toradex/requirements.txt
pyserial
gpiod